+++ src/blocks/notification/Notification.types.ts
@@ -1,8 +1,15 @@
 import { ReactNode } from 'react';
 
 export type NotificationProps = {
+  image?: ReactNode;
+  /* Title of the notification */
+  title?: string;
+  /* Description of the notification */
+  description?: string;
+  /* Optional onClose action for the notification */
+  onClose?: () => void;
   /* Custom React component to be passed as the image. */
-  overlay: ReactNode;
+  overlay?: ReactNode;
   /* Optional onClick event for the notification */
   onClick?: () => void;
   /* Position of the notification */
