+++ src/common/Common.utils.tsx
@@ -1,8 +1,5 @@
-import { NotificationEvent } from '@pushprotocol/restapi';
 import { LOGO_ALIAS_CHAIN } from './Common.constants';
 import { networkName } from 'helpers/UtilityHelper';
-import { NotificationToast } from './components';
-import { notification } from 'blocks';
 
 export const getSelectChains = (chainIdList: Array<number>) => {
   return chainIdList?.map((key: number) => {
@@ -13,13 +10,4 @@ export const getSelectChains = (chainIdList: Array<number>) => {
       icon: <Component />,
     };
   });
-};
-
-export const handleNotificationToast = (data: NotificationEvent) => {
-  notification.show({
-    overlay: <NotificationToast notification={data} />,
-  });
-  // setTimeout(() => {
-  //   notification.hide();
-  // }, 10000);
 };
