+++ src/common/components/index.ts
@@ -5,5 +5,4 @@ export * from './UnsubscribeChannelDropdown';
 export * from './ModalHeader';
 export * from './StakingVariant';
 export * from './TokenFaucet';
-export * from './NotificationOverlay';
-export * from './NotificationToast';
+export * from './InAppNotifications';
