+++ src/common/hooks/index.ts
@@ -1,4 +1,4 @@
 export { useIsVisible } from './useIsVisible';
 export { usePushStakingStats } from './usePushStakingStats';
 export { useDisclosure } from './useDisclosure';
-export { useStream } from './useStream';
+export { useInAppNotifications } from './useInAppNotifications';
